extract GFlowsLib to separate file
